requirements.txt
future
mxnet
Pillow
glob
cv2
csv
scipy